Create POST endpoint for moods
This also has some user-related code in it as well, mainly because
SQLLite that I'm using for testing doesn't do well with altering
constraints (like foreign keys) after a table has already been created,
so I put them both in here to start.  Also, there's no validation of the
data coming in to the POST endpoint, so that would be something to
improve in the future.