Added ability to track streaks
Currently, we'll still show the last streak even if it's been days since
the last posted mood, the streak will only reset once you add a new
mood, per the description of the story. I would probably try to talk to
whoever was giving me these requirements to see if that was actually the
behavior intended, or if they'd want the streak to reset to zero on the
day after the first day with no posted mood.  Also, added some code for
handling getting moods from a user with no moods so far.