Show longest streak percentile in moods endpoint
This will add a "percentile" field to the body of the moods endpoint if
the user's longest streak is at least in the top 50 percentile.  I
decided to have this be a full integer and be rounded down, since that
agrees with the way I usually see percentiles showing up (e.g. 99.5
percentile would be in the 99th percentile). Though admittedly, I'm not
a statistician, and this might be an off-by-one error.  Also, fixed a
bug where I was calculating streaks based on the last posted mood, not
based on the currrent user's last posted mood.